fix(07-04): route loan-state query handlers through IAssetQueries
- Move snapshot timeline/latest-state read logic to IAssetQueries implementation
- Query handlers now depend on IAssetQueries instead of IAssetRepository
- Resolves QueryHandlerDependencyTests architecture violation (query handlers must not depend on IRepository)
- Updates handler tests to inject IAssetQueries
